+++ api/src/main/java/org/allaymc/api/world/Dimension.java
@@ -627,7 +627,7 @@ public interface Dimension {
         var zIndex = z & 15;
         var oldBlockState = chunk.getBlockState(xIndex, y, zIndex, layer);
 
-        var newBlockState = oldBlockState.setProperty(propertyType, value);
+        var newBlockState = oldBlockState.setPropertyValue(propertyType, value);
         if (oldBlockState == newBlockState) return;
 
         chunk.setBlockState(xIndex, y, zIndex, newBlockState, layer);