+++ server/src/main/java/org/allaymc/server/block/component/trapdoor/BlockTrapdoorBaseComponentImpl.java
@@ -43,11 +43,11 @@ public class BlockTrapdoorBaseComponentImpl extends BlockBaseComponentImpl {
         }
 
         var face = placementInfo.player().getHorizontalFace();
-        blockState = blockState.setProperty(DIRECTION_4, TRAPDOOR_DIRECTION.get(face));
+        blockState = blockState.setPropertyValue(DIRECTION_4, TRAPDOOR_DIRECTION.get(face));
 
         var blockFace = placementInfo.blockFace();
         if ((placementInfo.clickedPos().y() > 0.5 && blockFace != BlockFace.UP) || blockFace == BlockFace.DOWN) {
-            blockState = blockState.setProperty(BlockPropertyTypes.UPSIDE_DOWN_BIT, true);
+            blockState = blockState.setPropertyValue(BlockPropertyTypes.UPSIDE_DOWN_BIT, true);
         }
 
         dimension.setBlockState(placeBlockPos.x(), placeBlockPos.y(), placeBlockPos.z(), blockState, placementInfo);